I think this is basically done
